3619: Rename LongAttributeNameIssueGenerator files
