+++ common/src/View/MapDocument.cpp
@@ -59,8 +59,8 @@
 #include "Model/LinkSourceIssueGenerator.h"
 #include "Model/LinkTargetIssueGenerator.h"
 #include "Model/LockState.h"
-#include "Model/LongAttributeValueIssueGenerator.h"
 #include "Model/LongPropertyKeyIssueGenerator.h"
+#include "Model/LongPropertyValueIssueGenerator.h"
 #include "Model/MissingClassnameIssueGenerator.h"
 #include "Model/MissingDefinitionIssueGenerator.h"
 #include "Model/MissingModIssueGenerator.h"
